requirements.txt
spotipy
psycopg2-binary
python-dotenv
requests
fastapi
uvicorn
jinja2